requirements.txt
django
pandas
numpy
scikit-learn
joblib
djangorestframework
streamlit
psycopg2
gunicorn
whitenoise